Controller, Service, Repo Connection #1
Began connecting the three.

Changed the return type of fetchAll() to List<Adoption> from List<Dog>. The Dog object type may not be used and we need to return all Adoptions, which may contain Dogs.

Separate classes and interfaces should be implemented related to dogs if we use the Dog object type.